Extract l10n stuff into its own file
